medical_record_service/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
python-consul==1.1.0
requests==2.31.0
djangorestframework-simplejwt==5.3.0